refactor: :recycle: update `header/nav-bar.jsx` name to `header/index.jsx` in accordance with standards
updated link path for `NavBar` in `layout.jsx`
